+++ .pre-commit-config.yaml
@@ -17,7 +17,7 @@ repos:
       - id: detect-aws-credentials
         args: [--allow-missing-credentials]
       - id: detect-private-key
-        exclude: revengine/tests/test_settings.py
+        exclude: revengine/tests/test_utils.py
       - id: end-of-file-fixer
         exclude: revengine/static/apple-developer-merchantid-domain-association
       - id: trailing-whitespace
